refactor(cache): renames revision-data git-revision-data
