+++ src/main/index.js
@@ -7,7 +7,7 @@ const cruiseResultSchema = require("../schema/cruise-result.schema.js");
 const meta = require("../extract/transpile/meta");
 const bus = require("../utl/bus");
 const { canServeFromCache, readCache, writeCache } = require("../cache/cache");
-const { getRevisionData } = require("../cache/revision-data");
+const { getRevisionData } = require("../cache/git-revision-data");
 const normalizeFilesAndDirectories = require("./files-and-dirs/normalize");
 const validateRuleSet = require("./rule-set/validate");
 const normalizeRuleSet = require("./rule-set/normalize");
